getAuctionElements
